messages are mistakenly ignored by jabber-plugin
Nowdays non-xhtml mesages from clients ready to receive events
relating to the delivery, display, and composition of messages
are erroneously ignored by jabber plugin.

According to http://xmpp.org/extensions/xep-0022.html (Section 3.1)
if message has body and any (or all) of isOffline(), isDelivered(),
isComposing(), isDisplayed() flags set to `true', that means that
the sender wants to be notified if the message is stored offline
(by the server), notified when the message is delivered (to the
client), notified if the recipient begins replying to the message
and notified when the message is displayed to recipient.

In case of presence of body element these flags are absolutely
unrelated to current state of given message. Additional checks
should not be performed.