Bump to Smack 4.1.9
This is a first step towards improving the security of the
plugin. Smack 4.1.9 fixes CVE-2016-10027, but we need to configure
Smack so that it uses 'SecurityMode.required' if the users ticks a
"Use TLS" checkbox (which is yet to be created) when configuring the
plugin. Also instead of "Accept all SSL/TLS certificates" java-pinning
should be used.

Fixes JENKINS-45599.

Thanks to Friso Vrolijken for working together with me on this.